feat: Add `sortNodes` and generic `orderBy`
Move the functions from the files app to this library to allow
reusing them in other places, like for example the FilePicker or Viewer app.

`orderBy` is a generic natural sorting function which allows to pass
a collection, identifiers and associated order and sorts the
collectiong using the browser native `Intl.Collator`.

`sortNodes` is a specialized function based on `orderBy`
to sort nodes (files and folders) based on sorting options.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>