feat: Export public interfaces of `Node`, `File` and `Folder`
If you using those classes in Vue (e.g. in data) they will cause errors,
as Vue unrefs data deeply and this causes all private fields to be stripped off from the interface.
Passing one of those classes from data to a function that expects e.g. `Node` will then
cause a Typescript error because the passed value is lacking the private fields.

So instead this provives interfaces that can be used whenever a parameter should
be of one of those types.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>