fix: `Node.move` should also adjust the displayname
The displayname is always set by Nextcloud,
so if it is not a custom one it is set to the basename.
This means on move (and rename) we need to update the displayname.

This is an alternative to https://github.com/nextcloud/server/pull/46474

We should consider making the displayname a top-level attribute like basename.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>