fix: Make `View` methods bound to `ViewData` object
Use case:
You want to scope logic and state into the view rather than somewhere in a module.

Problem:
The `View` class wraps the `ViewData` as a getter,
meaning instead of calling the function on the object,
it just returns the reference to the function within the object.
This means the function looses the reference to the object as `this`.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>