fix: Make types of File and Folder more explicit
Both inherit from Node which is generic, thus e.g. the `mime` field is of type `string`.
But for folders the mime can never be something different to
`'httpd/unix-directory'`.
So the correct type would be the string literal
`'httpd/unix-directory'`.

The same for the `type` attribute for `File` and `Folder`.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>