refactor: use interfaces where possible instead of instances
This allows to use the methods in more places, especially because when
wrapped in e.g. Vue the signature of private fields will be changed.
Meaning that a class instance within a `ref` will not have the same
type due to its private fields.

But we do not need real instances here, instead just allow using the
public interfaces as they contain the public API.
This also allows us to have custom classes that e.g. implement the
`IView` but still we ensure that it satisfies our interface.

Using interfaces for API is not a breaking change, because it widenes
the function interface.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>